Move payment setting to organizer setting section
