+++ backend/config/app.php
@@ -47,8 +47,8 @@ return [
         'reset_password' => '/auth/reset-password/%s',
         'confirm_email_change' => '/manage/profile/confirm-email-change/%s',
         'accept_invitation' => '/auth/accept-invitation/%s',
-        'stripe_connect_return_url' => '/manage/organizer/%d/payments',
-        'stripe_connect_refresh_url' => '/manage/organizer/%d/payments',
+        'stripe_connect_return_url' => '/manage/organizer/%d/settings#payouts',
+        'stripe_connect_refresh_url' => '/manage/organizer/%d/settings#payouts',
         'event_homepage' => '/event/%d/%s',
         'attendee_product' => '/product/%d/%s',
         'order_summary' => '/checkout/%d/%s/summary',
